requirements.txt
aniso8601==7.0.0
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pytz==2019.2
six==1.12.0
Werkzeug==0.15.5
flask-sqlalchemy==2.4.0
cryptography==3.4.7
google-generativeai

